requirements.txt
matplotlib
numpy
opencv-python
pandas
pillow
streamlit
torch
torchvision
